+++ CHANGELOG.rst
@@ -15,6 +15,8 @@ Next Release (TBD)
   (`#1100 <https://github.com/aws/chalice/issues/1100>`__)
 * Update ``policies.json`` file
   (`#1110 <https://github.com/aws/chalice/issues/1110>`__)
+* Support repeating values in the query string
+  (`#1131 <https://github.com/aws/chalice/issues/1131>`__)
 
 
 1.8.0